+++ sugar/lib/sugar.dart
@@ -15,7 +15,7 @@ export 'collection.dart';
 export 'collection_aggregate.dart';
 export 'core.dart';
 export 'core_range.dart';
-export 'core_runtime.dart';
+export 'core_system.dart';
 export 'math.dart';
 export 'time.dart';
 export 'time_interop.dart';